backend/test-auth/requirements.txt
aiohttp>=3.9.0
livekit-api>=0.6.0
python-dotenv>=1.0.0